Update queryTwitterAPI.py
I found that the option to query the Twitter API and exclude paid content is not included in my student-level plan, so I adjusted my query to search for tweets to determine employee sentiment instead of consumer sentiment of a given company or brand. I left the brand sentiment as an option. If you have a premium or enterprise tier key to Twitter you can add -is:nullcast to remove tweets for promotion from ads.twitter